Documentation, Refractors, and PIP Updates
I have updated the PIP installer, removing bloat, and updating metadata associated with setup.cfg. This includes moving /examples/ under the /docs/ folder to remove clutter in the root directory. In addition to, updating project dependencies (avoiding conflicts with past code) -- hence, new versions of the codebase will not be backwards-compatible as of now. Documentation and README.md has been updated to reflect the changes